+++ pages/api/admin/index.js
@@ -1,0 +1,3 @@
+export default async function handler(req, res) {
+  res.json("/api/admin");
+}
