+++ pages/api/admin/post/[id].js
@@ -1,33 +1,34 @@
-import prisma from "../../../prisma/client";
+import prisma from "../../../../prisma/client";
 
 export default async function handler(req, res) {
-  //get, put para post por id
   const { id } = req.query;
   if (req.method === "GET") {
+    // Obtener todos los posts
     const post = await prisma.post.findUnique({
       where: {
         id: id,
       },
       include: {
+        author: true,
         reviews: true,
       },
     });
     res.status(200).json(post);
   } else if (req.method === "PUT") {
-    const { title, content, photo, category, price, type } = req.body;
     const post = await prisma.post.update({
       where: {
         id: id,
       },
+      include: {
+        author: true,
+        reviews: true,
+      },
       data: {
-        title,
-        content,
-        photo,
-        category,
-        price,
-        type,
+        ...req.body,
       },
     });
     res.status(200).json(post);
+  } else {
+    res.status(405).json({ message: "Método HTTP no permitido" });
   }
 }
