+++ pages/api/order/[id].js
@@ -22,29 +22,6 @@ export default async function handler(req, res) {
       console.error(error);
       res.status(500).json({ message: "Error retrieving order" });
     }
-  } else if (req.method === "PUT") {
-    const { status, userId, postId, paymentId } = req.body;
-
-    try {
-      const updatedOrder = await prisma.order.update({
-        where: { id },
-        data: {
-          status,
-          userId,
-          postId,
-          paymentId,
-        },
-        include: {
-          payment: true,
-          user: true,
-        },
-      });
-
-      res.status(200).json(updatedOrder);
-    } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Error updating order" });
-    }
   } else {
     res.status(405).json({ message: "Method not allowed" });
   }
