+++ pages/api/order/index.js
@@ -1,20 +1,7 @@
 import prisma from "../../../prisma/client";
 
 export default async function handler(req, res) {
-  if (req.method === "GET") {
-    try {
-      const orders = await prisma.order.findMany({
-        include: {
-          payment: true,
-          user: true,
-        },
-      });
-
-      res.status(200).json(orders);
-    } catch (error) {
-      res.status(500).json({ message: "Error retrieving orders" });
-    }
-  } else if (req.method === "POST") {
+  if (req.method === "POST") {
     const { status, userId, postId, paymentId } = req.body;
 
     try {
