+++ pages/api/payment/[id].js
@@ -18,27 +18,6 @@ export default async function handler(req, res) {
     } catch (error) {
       res.status(500).json({ message: "Error retrieving payment" });
     }
-  } else if (req.method === "PUT") {
-    const { id } = req.query;
-    const { amount, currency, userId } = req.body;
-    try {
-      const payment = await prisma.payment.update({
-        where: {
-          id: id,
-        },
-        data: {
-          amount,
-          currency,
-          userId,
-        },
-        include: {
-          user: true,
-        },
-      });
-      res.status(200).json(payment);
-    } catch (error) {
-      res.status(500).json({ message: "Error updating payment" });
-    }
   } else {
     res.status(405).json({ message: "Method not allowed" });
   }
