+++ pages/api/user/[id].js
@@ -1,6 +1,6 @@
 import { parse } from "postcss";
 import prisma from "../../../prisma/client";
-const bcrypt = require('bcryptjs')
+const bcrypt = require("bcryptjs");
 
 export default async function handler(req, res) {
   const { id } = req.query;
@@ -18,39 +18,23 @@ export default async function handler(req, res) {
     });
     res.status(200).json(user);
   } else if (req.method === "PUT") {
-    const {
-      firstname,
-      lastname,
-      email,
-      admin,
-      logged,
-      hidden,
-      phoneNumber,
-      zipCode,
-      reports,
-    } = req.body;
-    let { 
-      password 
-    } = req.body
-    const rondasDeSal = 10
-    password = await bcrypt.hash(password, rondasDeSal)
-    const user = await prisma.user.update({
-      where: {
-        email: id,
-      },
-      data: {
-        firstname,
-        lastname,
-        email,
-        password,
-        admin,
-        logged,
-        hidden,
-        phoneNumber,
-        zipCode,
-        reports,
-      },
-    });
-    res.status(200).json(user);
+    const { firstname, lastname, password, phoneNumber, zipCode } = req.body;
+    try {
+      const user = await prisma.user.update({
+        where: {
+          email: id,
+        },
+        data: {
+          firstname,
+          lastname,
+          password,
+          phoneNumber,
+          zipCode,
+        },
+      });
+      res.status(200).json(user);
+    } catch (error) {
+      res.status(500).json({ error: "Error updating user." });
+    }
   }
 }
